fix: improve API compliance and input validation for REST endpoints
- Add @ApiResponse annotations for missing/undocumented HTTP status codes
  across multiple resource classes (ExternalImageResource,
  RecordsFilesImageResource, CMSMediaImageResource, and others)
- Add input validation for PI, Solr field names, date patterns, and
  image URLs to prevent HTTP 500 from illegal URI/Solr characters
- Fix ExternalImageResource URL decoding (URI.getPath() instead of
  URLDecoder to correctly handle '+' in paths)
- Add fallback GET handler in CMSMediaResource returning 400 for
  non-image filenames instead of JAX-RS 405
- Add ClientApplicationUpdate DTO for PUT /clients/{id} to avoid
  schemathesis readOnly+example schema conflict
- Fix AuthenticationEndpoint to read form params via
  HttpServletRequest.getParameter() to avoid IllegalStateException
- Add @JsonIgnoreProperties(ignoreUnknown=true) to
  RecordsRequestParameters and BookmarkList
- Correct OpenAPI schema types in ClientApplication (int64, date-time)
- Add unit tests for validateImageUrl(), requireValidListId(),
  and Solr field name regex validation

+++ goobi-viewer-core/src/main/java/io/goobi/viewer/api/rest/resourcebuilders/AnnotationsResourceBuilder.java
@@ -111,7 +111,9 @@ public class AnnotationsResourceBuilder {
         if (page == null || page < 1) {
             throw new IllegalRequestException("Page number must be at least 1");
         }
-        int first = (page - 1) * MAX_ANNOTATIONS_PER_PAGE;
+        // Use long arithmetic to prevent integer overflow when page is large (e.g. Integer.MAX_VALUE).
+        // Stream.skip() accepts long, so this is safe.
+        long first = (long)(page - 1) * MAX_ANNOTATIONS_PER_PAGE;
         String sortField = "id";
 
         List<IAnnotation> annotations = DataManager.getInstance()
